requirements.txt
Django~=3.0.6
wheel
django-crontab==0.7.1
django-taggit==1.2.0
django-crispy-forms